minor: validate mypy pre-commit

+++ pandora/matching_cost/sad_ssd.py
@@ -144,7 +144,7 @@ class SadSsd(matching_cost.AbstractMatchingCost):
         if self._subpix == 1:
             disparity_range = np.arange(disp_min, disp_max + 1)
         else:
-            disparity_range = np.arange(disp_min, disp_max, step=1 / float(self._subpix))
+            disparity_range = np.arange(disp_min, disp_max, step=1 / float(self._subpix), dtype=np.float64)
             disparity_range = np.append(disparity_range, [disp_max])
 
         # Allocate the numpy cost volume cv = (disp, col, row), for efficient memory management
